Revert "Abstract tests into more logical components"
