+++ tests/test-bylines-object.php
@@ -13,77 +13,6 @@ use Bylines\Objects\Byline;
 class Test_Bylines_Object extends WP_UnitTestCase {
 
 	/**
-	 * Create a new byline from thin air
-	 */
-	public function test_create_byline() {
-		$byline = Byline::create( array(
-			'display_name' => 'Foo Bar',
-			'slug'         => 'foobar',
-		) );
-		$this->assertInstanceOf( 'Bylines\Objects\Byline', $byline );
-		$this->assertEquals( 'Foo Bar', $byline->display_name );
-		$this->assertEquals( 'foobar', $byline->slug );
-	}
-
-	/**
-	 * Creating a byline but missing arguments
-	 */
-	public function test_create_byline_missing_arguments() {
-		$byline = Byline::create( array() );
-		$this->assertInstanceOf( 'WP_Error', $byline );
-		$this->assertEquals( 'missing-slug', $byline->get_error_code() );
-		$byline = Byline::create( array(
-			'slug' => 'foobar',
-		) );
-		$this->assertInstanceOf( 'WP_Error', $byline );
-		$this->assertEquals( 'missing-display_name', $byline->get_error_code() );
-	}
-
-	/**
-	 * Creating a byline from an existing user
-	 */
-	public function test_create_byline_from_user() {
-		$user_id = $this->factory->user->create( array(
-			'display_name'    => 'Foo Bar',
-			'first_name'      => 'Foo',
-			'last_name'       => 'Bar',
-			'user_email'      => 'foobar@gmail.com',
-			'user_login'      => 'foobar',
-		) );
-		$byline = Byline::create_from_user( $user_id );
-		$this->assertInstanceOf( 'Bylines\Objects\Byline', $byline );
-		$this->assertEquals( $user_id, $byline->user_id );
-		$this->assertEquals( 'Foo Bar', $byline->display_name );
-		$this->assertEquals( 'foobar', $byline->slug );
-		$this->assertEquals( 'Foo', $byline->first_name );
-		$this->assertEquals( 'Bar', $byline->last_name );
-		$this->assertEquals( 'foobar@gmail.com', $byline->user_email );
-		$this->assertEquals( 'foobar', $byline->user_login );
-	}
-
-	/**
-	 * Creating a byline from a user that doesn't exist
-	 */
-	public function test_create_byline_from_missing_user() {
-		$byline = Byline::create_from_user( BYLINES_IMPOSSIBLY_HIGH_NUMBER );
-		$this->assertInstanceOf( 'WP_Error', $byline );
-		$this->assertEquals( 'missing-user', $byline->get_error_code() );
-	}
-
-	/**
-	 * Creating a byline from a user that already has a byline
-	 */
-	public function test_create_byline_from_existing_user_byline() {
-		$user_id = $this->factory->user->create();
-		// Create the first byline.
-		Byline::create_from_user( $user_id );
-		// Attempt creating the second byline.
-		$byline = Byline::create_from_user( $user_id );
-		$this->assertInstanceOf( 'WP_Error', $byline );
-		$this->assertEquals( 'existing-byline', $byline->get_error_code() );
-	}
-
-	/**
 	 * Bylines object should look roughly like a WP_User object
 	 */
 	public function test_bylines_object_looks_like_wp_user_object() {
